update generate messagerecord move function to hook folder
