+++ package.json
@@ -13,6 +13,7 @@
     "@emotion/styled": "^11.11.0",
     "@mui/icons-material": "^5.14.16",
     "@mui/material": "^5.14.17",
+    "@mui/styles": "^5.14.18",
     "@react-oauth/google": "^0.12.1",
     "animate.css": "^4.1.1",
     "axios": "^1.6.2",
