+++ src/pages/createDocuments/paperflow.tsx
@@ -1,12 +1,11 @@
 import Image from "next/image";
 import styles from "./paperFlow.module.css";
 import { NextRouter, useRouter } from "next/router";
-
 // TODO : This function should be moved to a separate file
 const NavigatToPages = (router: NextRouter, path: string) => {
   switch (path) {
     case 'record':
-      router.push('/createDocuments/createpage1')
+      router.push('/createDocuments/pages/messageRecord')
       break;
     case 'please':
       router.push('')
